requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
Pillow==12.0.0
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
opencv-python==4.8.1.78
scikit-learn==1.4.0
